+++ .pre-commit-config.yaml
@@ -13,7 +13,7 @@ repos:
     rev: v0.7.3
     hooks:
       - id: ruff
-        args: [--fix, --exit-non-zero-on-fix]
+        args: [--fix, --select, I, --exit-non-zero-on-fix]
       - id: ruff-format
   - repo: https://github.com/pre-commit/mirrors-mypy
     rev: v1.13.0
